Move `RsPsiSubstitution.toSubst` to `TyLowering`
