+++ src/main/kotlin/org/rust/lang/core/types/RsPsiSubstitution.kt
@@ -5,20 +5,9 @@
 
 package org.rust.lang.core.types
 
-import com.intellij.openapi.util.RecursionGuard
-import com.intellij.openapi.util.RecursionManager
-import com.intellij.psi.PsiElement
 import org.rust.lang.core.psi.*
 import org.rust.lang.core.psi.ext.RsElement
-import org.rust.lang.core.types.consts.CtConstParameter
-import org.rust.lang.core.types.consts.CtUnknown
-import org.rust.lang.core.types.infer.TyLowering
-import org.rust.lang.core.types.infer.resolve
-import org.rust.lang.core.types.infer.substitute
-import org.rust.lang.core.types.regions.ReEarlyBound
-import org.rust.lang.core.types.ty.*
-import org.rust.lang.utils.evaluation.PathExprResolver
-import org.rust.lang.utils.evaluation.toConst
+import org.rust.lang.core.types.ty.Ty
 
 /** Similar to [Substitution], but maps PSI to PSI instead of [Ty] to [Ty] */
 open class RsPsiSubstitution(
@@ -45,70 +34,3 @@ open class RsPsiSubstitution(
         object FnSugarImplicitRet : AssocValue()
     }
 }
-
-fun RsPsiSubstitution.toSubst(
-    resolver: PathExprResolver,
-    tyLowering: TyLowering,
-): Substitution {
-    val typeSubst = typeSubst.entries.associate { (param, value) ->
-        val paramTy = TyTypeParameter.named(param)
-        val valueTy = when (value) {
-            is RsPsiSubstitution.Value.DefaultValue -> {
-                val defaultValue = value.value.value
-                val defaultValueTy = guard.doPreventingRecursion(defaultValue, /* memoize = */true) {
-                    defaultValue.rawType
-                } ?: TyUnknown
-                if (value.value.selfTy != null) {
-                    defaultValueTy.substitute(mapOf(TyTypeParameter.self() to value.value.selfTy).toTypeSubst())
-                } else {
-                    defaultValueTy
-                }
-            }
-            is RsPsiSubstitution.Value.OptionalAbsent -> paramTy
-            is RsPsiSubstitution.Value.Present -> when (value.value) {
-                is RsPsiSubstitution.TypeValue.InAngles -> tyLowering.lowerTy(value.value.value)
-                is RsPsiSubstitution.TypeValue.FnSugar -> if (value.value.inputArgs.isNotEmpty()) {
-                    TyTuple(value.value.inputArgs.map { if (it != null) tyLowering.lowerTy(it) else TyUnknown })
-                } else {
-                    TyUnit.INSTANCE
-                }
-            }
-            RsPsiSubstitution.Value.RequiredAbsent -> TyUnknown
-        }
-        paramTy to valueTy
-    }
-
-    val regionSubst = regionSubst.entries.mapNotNull { (psiParam, psiValue) ->
-        val param = ReEarlyBound(psiParam)
-        val value = if (psiValue is RsPsiSubstitution.Value.Present) {
-            psiValue.value.resolve()
-        } else {
-            return@mapNotNull null
-        }
-
-        param to value
-    }.toMap()
-
-    val constSubst = constSubst.entries.associate { (psiParam, psiValue) ->
-        val param = CtConstParameter(psiParam)
-        val value = when (psiValue) {
-            RsPsiSubstitution.Value.OptionalAbsent -> param
-            RsPsiSubstitution.Value.RequiredAbsent -> CtUnknown
-            is RsPsiSubstitution.Value.Present -> {
-                val expectedTy = psiParam.typeReference?.normType ?: TyUnknown
-                psiValue.value.toConst(expectedTy, resolver)
-            }
-            is RsPsiSubstitution.Value.DefaultValue -> {
-                val expectedTy = psiParam.typeReference?.normType ?: TyUnknown
-                psiValue.value.toConst(expectedTy, resolver)
-            }
-        }
-
-        param to value
-    }
-
-    return Substitution(typeSubst, regionSubst, constSubst)
-}
-
-private val guard: RecursionGuard<PsiElement> =
-    RecursionManager.createGuard("org.rust.lang.core.types.RsPsiSubstitution")
